[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17636 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,9 +4,9 @@ public class Main {
   public static void main(String[] args){
     Scanner scan = new Scanner(System.in);
     String [] nums = scan.nextLine().split(" ");
-    int num1 = Integer.parseInt(nums[0]);
-    int num2 = Integer.parseInt(nums[1]);
-    int result = (num1+num2) * (num1-num2);
+    long num1 = Integer.parseInt(nums[0]);
+    long num2 = Integer.parseInt(nums[1]);
+    long result = (num1+num2) * (num1-num2);
     System.out.println(result);
   }
 }
